pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openrouter>=0.2.1",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
]
